# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Gestion de biens immobilier",
        "description": "Documentation de l'API immoApp",
        "version": "1.0.0"
    },
    "paths": {
        "/api/avis": {
            "get": {
                "tags": [
                    "Avis"
                ],
                "summary": "Liste tous les avis",
                "operationId": "074b394984977e6dba6974626b17ff22",
                "responses": {
                    "200": {
                        "description": "Succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Avis"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Avis"
                ],
                "summary": "Créer un avis",
                "operationId": "c6eeaa4be5353dc347263747d29b964d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Avis"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Avis créé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Avis"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/categorie_biens": {
            "get": {
                "tags": [
                    "CategorieBiens"
                ],
                "summary": "Lister catégories",
                "operationId": "f2a4fa0d328368d571f467308348926c",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "tags": [
                    "CategorieBiens"
                ],
                "summary": "Créer catégorie",
                "operationId": "10630faa2dba82bb3f1a237fd116e47c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Créé"
                    }
                }
            }
        },
...
        "/contrats": {
            "get": {
                "tags": [
                    "Contrats"
                ],
                "summary": "Liste tous les contrats",
                "operationId": "53cd237a004976e277cf953b0b8382e4",
                "responses": {
                    "200": {
                        "description": "Succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Contrats"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Contrats"
                ],
                "summary": "Créer un contrat",
                "operationId": "903df457109981b6d3ab8bb04cdef859",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Contrats"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Contrat créé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Contrats"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/payements": {
            "get": {
                "tags": [
                    "Payements"
                ],
                "summary": "Liste tous les payements",
                "operationId": "4d17b6fc5b3c353934034c9a0082d30e",
                "responses": {
                    "200": {
                        "description": "Succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Payements"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Payements"
                ],
                "summary": "Créer un payement",
                "operationId": "c46eacdb787d3d9e40763749a0d0cddb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Payements"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Payement créé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Payements"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/photos": {
            "get": {
                "tags": [
                    "Photos"
                ],
                "summary": "Liste toutes les photos",
                "operationId": "d34843ba16a0674d6f5da4274b3f5b23",
                "responses": {
                    "200": {
                        "description": "Succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Photos"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Photos"
                ],
                "summary": "Créer une photo",
                "operationId": "3a21edcca99bb9e6660ee5da3df6d317",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Photos"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Photo créée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Photos"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/quartiers": {
            "get": {
                "tags": [
                    "Quartiers"
                ],
                "summary": "Liste tous les quartiers",
                "operationId": "6b0fde1ea0c688bfa7621d4a465a5aab",
                "responses": {
                    "200": {
                        "description": "Succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Quartiers"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Quartiers"
                ],
                "summary": "Créer un quartier",
                "operationId": "c2bf53c8080ef2a6a756b07aeda5bce7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Quartiers"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Quartier créé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Quartiers"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/reservations": {
            "get": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Liste toutes les réservations",
                "operationId": "ac05133c4d78a4319a5f091fa1686b3b",
                "responses": {
                    "200": {
                        "description": "Succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Reservations"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Créer une réservation",
                "operationId": "180a4e6945149199996c45b5a6c9744e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Reservations"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Réservation créée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Reservations"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/type_biens": {
            "get": {
                "tags": [
                    "TypeBiens"
                ],
                "summary": "Lister types",
                "operationId": "2fcad4a252fd4481968e525c3f3fdb21",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "tags": [
                    "TypeBiens"
                ],
                "summary": "Créer type",
                "operationId": "f4e6af76ad921623afa4719ce89946ea",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Créé"
                    }
                }
            }
        },
        "/api/type_biens/{id}": {
            "get": {
                "tags": [
                    "TypeBiens"
                ],
                "summary": "Afficher type",
                "operationId": "29e0c00758a5d75d3df7338f659f57a6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails"
                    }
                }
            },
            "put": {
                "tags": [
                    "TypeBiens"
                ],
                "summary": "Modifier type",
                "operationId": "a986a50c22fb0bc92821aecf3520ca16",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Modifié"
                    }
                }
            },
            "delete": {
                "tags": [
                    "TypeBiens"
                ],
                "summary": "Supprimer type",
                "operationId": "45342a7d5396276b5889b914c56c0e66",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Supprimé"
                    }
                }
            }
        },
...
        "/utilisateurs": {
            "get": {
                "tags": [
                    "Utilisateurs"
                ],
                "summary": "Liste tous les utilisateurs",
                "operationId": "3ca4091dbd6454ea1ec2eb43f5db6c3b",
                "responses": {
                    "200": {
                        "description": "Succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Utilisateurs"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Utilisateurs"
                ],
                "summary": "Créer un utilisateur",
                "operationId": "01edc46161789482f77838a996fd3943",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Utilisateurs"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Utilisateur créé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Utilisateurs"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/villes": {
            "get": {
                "tags": [
                    "Villes"
                ],
                "summary": "Lister villes",
                "operationId": "45da93baea60a6e346e095f1f1569e6b",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "tags": [
                    "Villes"
                ],
                "summary": "Créer ville",
                "operationId": "1fa6a9eaa07969c49560b1d81c194187",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Créé"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Avis": {
                "required": [
                    "utilisateur_id",
                    "commentaire",
                    "note"
                ],
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "utilisateur_id": {
                        "type": "integer"
                    },
                    "commentaire": {
                        "type": "string"
                    },
                    "note": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "Contrats": {
                "required": [
                    "description"
                ],
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "description": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "Payements": {
                "required": [
                    "montant",
                    "type_payement_id",
                    "reservation_id"
                ],
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "montant": {
                        "type": "number",
                        "format": "float"
                    },
                    "type_payement_id": {
                        "type": "integer"
                    },
                    "reservation_id": {
                        "type": "integer"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Photos": {
                "required": [
                    "url"
                ],
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "url": {
                        "type": "string"
                    },
                    "bien_id": {
                        "type": "integer"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Quartiers": {
                "required": [
                    "nom",
                    "ville_id"
                ],
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "nom": {
                        "type": "string",
                        "example": "Bonapriso"
                    },
                    "ville_id": {
                        "type": "integer"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Reservations": {
                "required": [
                    "utilisateur_id",
                    "contrat_id",
                    "date_debut",
                    "date_fin"
                ],
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "utilisateur_id": {
                        "type": "integer"
                    },
                    "contrat_id": {
                        "type": "integer"
                    },
                    "date_debut": {
                        "type": "string",
                        "format": "date"
                    },
                    "date_fin": {
                        "type": "string",
                        "format": "date"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "Utilisateurs": {
                "required": [
                    "nom",
                    "prenom",
                    "email",
                    "mot_de_passe",
                    "type_utilisateur_id"
                ],
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "nom": {
                        "type": "string"
                    },
                    "prenom": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "mot_de_passe": {
                        "type": "string"
                    },
                    "type_utilisateur_id": {
                        "type": "integer"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}